phase-d(bun_runtime): ByteBlobLoader: drop u32 cast on IntoArray.len (now BlobSizeType)

+++ src/runtime/webcore/ByteBlobLoader.rs
@@ -157,13 +157,13 @@ impl ByteBlobLoader {
         if self.remain == 0 {
             return streams::Result::IntoArrayAndDone(streams::IntoArray {
                 value: array,
-                len: copied as u32,
+                len: copied,
             });
         }
 
         streams::Result::IntoArray(streams::IntoArray {
             value: array,
-            len: copied as u32,
+            len: copied,
         })
     }
